feat: Add method-override dependency for handling HTTP methods
This commit adds the "method-override" npm package as a dependency in the package.json file. This package allows us to handle HTTP methods such as PUT and DELETE in forms by overriding the method using a query parameter or a hidden input field. This will be useful for implementing edit and delete functionality in our application.